Refactor login user mutation
