+++ package.json
@@ -8,8 +8,6 @@
     "heroku-postbuild": "babel src --out-dir dist --copy-files",
     "dev": "env-cmd -f ./config/dev.env nodemon src/index.js --ext js,graphql,env --exec babel-node",
     "get-schema": "env-cmd -f ./config/dev.env graphql get-schema -p prisma",
-    "ptoken": "prisma token -e ../config/dev.env",
-    "pdev": "prisma deploy -e ../config/dev.env",
     "test": "jest --watch"
   },
   "repository": {
