+++ src/resolvers/mutations/ExternalAccount.js
@@ -1,0 +1,3 @@
+export const ExternalAccount = {
+
+}
