+++ src/resolvers/mutations/index.js
@@ -1,7 +1,11 @@
+import { UserAccount } from './UserAccount';
+import { ExternalAccount } from './ExternalAccount'
 import { UserProfile } from './UserProfile';
 import { Project } from './Project';
 
 export const Mutation = {
+	...UserAccount,
+	...ExternalAccount,
 	...UserProfile,
 	...Project
 };
