+++ src/resolvers/types.js
@@ -1,5 +1,9 @@
 import { getUserId } from '../utils/getUserId';
 
+
+const UserAccount = {}
+const ExternalAccount = {}
+
 const UserProfile = {
 	email: {
 		fragment: 'fragment userId on User { id }',
@@ -19,5 +23,7 @@ const Project = {};
 
 export const Types = {
 	UserProfile,
+	UserAccount,
+	ExternalAccount,
 	Project
 };
